# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/projects/{project_name}/pipelines/{pipeline_name}/secrets": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "secrets"
                ],
                "summary": "Get secrets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pipeline name",
                        "name": "pipeline_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order by field",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Secret key (pattern)",
                        "name": "key",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of secrets",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Secret"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "secrets"
                ],
                "summary": "Create new secret",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pipeline name",
                        "name": "pipeline_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New secret entry",
                        "name": "secret",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SecretInput"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/projects/{project_name}/pipelines/{pipeline_name}/secrets/{secret_key}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "secrets"
                ],
                "summary": "Update secret",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pipeline name",
                        "name": "pipeline_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Secret key",
                        "name": "secret_key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated secret entry",
                        "name": "secret",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SecretInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated secret",
                        "schema": {
                            "$ref": "#/definitions/model.Secret"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "secrets"
                ],
                "summary": "Delete secret",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pipeline name",
                        "name": "pipeline_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Secret key",
                        "name": "secret_key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/projects/{project_name}/secrets": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "secrets"
                ],
                "summary": "Get secrets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order by field",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Secret key (pattern)",
                        "name": "key",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of secrets",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Secret"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "secrets"
                ],
                "summary": "Create new secret",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New secret entry",
                        "name": "secret",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SecretInput"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/projects/{project_name}/secrets/{secret_key}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "secrets"
                ],
                "summary": "Update secret",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Secret key",
                        "name": "secret_key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated secret entry",
                        "name": "secret",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SecretInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated secret",
                        "schema": {
                            "$ref": "#/definitions/model.Secret"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "secrets"
                ],
                "summary": "Delete secret",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Secret key",
                        "name": "secret_key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/secrets": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "secrets"
                ],
                "summary": "Get secrets",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order by field",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Secret key (pattern)",
                        "name": "key",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of secrets",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Secret"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "secrets"
                ],
                "summary": "Create new secret",
                "parameters": [
                    {
                        "description": "New secret entry",
                        "name": "secret",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SecretInput"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/secrets/{secret_key}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "secrets"
                ],
                "summary": "Update secret",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Secret key",
                        "name": "secret_key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated secret entry",
                        "name": "secret",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SecretInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated secret",
                        "schema": {
                            "$ref": "#/definitions/model.Secret"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "secrets"
                ],
                "summary": "Delete secret",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Secret key",
                        "name": "secret_key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Secret": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.SecretInput": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
    }
}